+++ Cargo.toml
@@ -5,18 +5,17 @@ authors = ["zer0 <matteo@dusk.network>", "Victor Lopez <victor@dusk.network"]
 edition = "2018"
 
 [dependencies]
-rand_core = "0.5"
+rand_core = "0.6"
 dusk-bytes = "0.1"
-dusk-bls12_381 = {version = "0.6", default-features = false}
-dusk-jubjub = {version = "0.8", default-features = false}
-dusk-poseidon = {version = "0.20", default-features = false}
-dusk-pki = {version = "0.6", default-features = false}
-canonical = {version = "0.5", optional = true}
-canonical_derive = {version = "0.5", optional = true}
+dusk-bls12_381 = { version = "0.8.0-rc", default-features = false }
+dusk-jubjub = { version = "0.10.0-rc", default-features = false }
+dusk-poseidon = { version = "0.21.0-rc", default-features = false }
+dusk-pki = { version = "0.7.0-rc", default-features = false }
+canonical = { version = "0.6", optional = true }
+canonical_derive = { version = "0.6", optional = true }
 
 [dev-dependencies]
 assert_matches = "1.3"
-rand = "0.7"
 
 [features]
 canon = [
